Move `NativeExecutorContext` into its own module
